[P2P-110] test: PlaceLikeService 테스트 오류 수정
